requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
psycopg2==2.7.2
